+++ tests/test_cmd2.py
@@ -1676,13 +1676,6 @@ def test_alias_create(base_app):
     out, err = run_cmd(base_app, 'alias list fake')
     assert out == normalize('alias create fake help')
 
-    # Test silent flag
-    out, err = run_cmd(base_app, 'alias create --silent fake set')
-    assert not out
-
-    out, err = run_cmd(base_app, 'alias list --with_silent fake')
-    assert out == normalize('alias create --silent fake set')
-
 def test_alias_create_with_quoted_tokens(base_app):
     """Demonstrate that quotes in alias value will be preserved"""
     create_command = 'alias create fake help ">" "out file.txt" ";"'
@@ -1785,13 +1778,6 @@ def test_macro_create(base_app):
     # Look up the updated macro
     out, err = run_cmd(base_app, 'macro list fake')
     assert out == normalize('macro create fake help')
-
-    # Test silent flag
-    out, err = run_cmd(base_app, 'macro create --silent fake set')
-    assert not out
-
-    out, err = run_cmd(base_app, 'macro list --with_silent fake')
-    assert out == normalize('macro create --silent fake set')
 
 def test_macro_create_with_quoted_tokens(base_app):
     """Demonstrate that quotes in macro value will be preserved"""